Star System UI has basic function. Will now look at a ship manager for fleets, move ships, from fleet to fleet, add ship...

+++ Assets/Script/Galactic/StarSysData.cs
@@ -23,7 +23,8 @@ public class StarSysData
     public int food;
     public int power;
     public int production;
-    public float tech;
+    public int tech;
+
     public string Description;
     private string v;
